Eliminate (most) "signed and unsigned comparison" warnings from agent code [-Wsign-compare]
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@19018 06827809-a52a-0410-b366-d66718629ded

+++ agent/helpers/row_merge.c
@@ -76,7 +76,7 @@ netsnmp_row_merge_status_get(netsnmp_handler_registration *reginfo,
      * see if we've already been here
      */
     rc = snprintf(buf, sizeof(buf), "row_merge:%p", reginfo);
-    if ((-1 == rc) || (rc >= sizeof(buf))) {
+    if ((-1 == rc) || ((size_t)rc >= sizeof(buf))) {
         snmp_log(LOG_ERR,"error creating key\n");
         return NULL;
     }
